fixed the linking issus
